Create client side component of alphabet app
This app has 2 phases, Lesson and Review. During the
lesson phase, the student is shown flash cards which when clicked
plays an audio of the sound a letter makes. The number of letters
tested will be decided by the back-end.

During the review phase, the student is required to select the
letter that matches the audio played.

The CSS needs to be finalised.